refactor(dialogs): restructure to repo layout
