--- apps/tissuumaps/src/components/dialogs/index.ts
@@ -1,8 +1,0 @@
-export { DialogProvider } from "./dialog-provider";
-export { useAlert, useConfirm, usePrompt } from "./dialog-context";
-export type {
-  AlertParams,
-  ConfirmParams,
-  DialogAction,
-  PromptParams,
-} from "./dialog-context";
