+++ apps/tissuumaps/src/components/panels/LabelsPanel/index.tsx
@@ -12,7 +12,7 @@ import {
   AccordionTrigger,
   AccordionTriggerUpDownIcon,
 } from "@/components/common/accordion";
-import { useConfirm } from "@/components/dialogs";
+import { useConfirmDialog } from "@/components/dialogs/ConfirmDialog/hooks";
 import { Button } from "@/components/ui/button";
 import {
   InputGroup,
@@ -101,7 +101,7 @@ function LabelsAccordionItem({ labels, index }: LabelsAccordionItemProps) {
 
   const updateLabels = useProjectStore((state) => state.updateLabels);
   const deleteLabels = useProjectStore((state) => state.deleteLabels);
-  const confirm = useConfirm();
+  const confirm = useConfirmDialog();
 
   const labelsData = useLabelsData(labels.id);
